backend/requirements.txt
flask==3.1.0
twilio==9.4.0
requests==2.32.3
python-dotenv==1.1.0
flask-socketio==5.5.1
pydub==0.25.1
pyngrok==7.2.2
